enh: convert purchasing money fields to Decimal (MONEY-4) (#19)
* enh: convert purchasing money fields to Decimal (MONEY-4)

Convert the 14 purchasing IDR money fields (PO totals, base-currency
item/discounted prices, shipping/delivery/commission fees) from
BigIntegerField to DecimalField(18,2). RMB decimal fields and
exchange_rate stay untouched. Update service coercions and serializer
output to preserve exact values; round only at report-serialization
boundaries and where writing into still-integer fields on other apps
(finance AP, inventory COGS) via round_money_to_int.

* refactor: extract shared DELIVERED inventory-data builder in purchasing_service

update_purchase_order built the same inventory_data payload for a DELIVERED
receipt in two places — the initial DELIVERED transition and a later
incremental received_qty update on an already-DELIVERED PO — differing only
in the delivery-date gate vs. baseline-delta computation. Extract
_build_delivery_inventory_data to share the common filtering and dict shape,
and drop the _ids2/_company_id2 naming residue from the copy-paste. Add a
characterization test for the incremental path, which had no prior coverage.

+++ apps/purchasing/services/sourcing_product_service.py
@@ -43,8 +43,8 @@ def _create_po_detail(
     unit_price_foreign = sourcing_row.unit_price
     discounted_price_foreign = sourcing_row.discounted_price or unit_price_foreign
 
-    unit_price_base = int(round(unit_price_foreign * exchange_rate))
-    discounted_unit_price_base = int(round(discounted_price_foreign * exchange_rate))
+    unit_price_base = Decimal(round(unit_price_foreign * exchange_rate))
+    discounted_unit_price_base = Decimal(round(discounted_price_foreign * exchange_rate))
 
     return PurchaseOrderDetail.objects.create(
         purchase_order=po,
